Move Asset Component to own files - updates to column alignments - Added use `createAccountSubString` to breakdown long NFT ids - use word-break to break up long names
