+++ src/components/pages/DAOTreasury/components/Assets.tsx
@@ -29,210 +29,9 @@ import EtherscanLink from '../../../ui/links/EtherscanLink';
 import { ModalType } from '../../../ui/modals/ModalProvider';
 import { useFractalModal } from '../../../ui/modals/useFractalModal';
 import Divider from '../../../ui/utils/Divider';
-import { TokenDisplayData, useFormatCoins } from '../hooks/useFormatCoins';
-
-function CoinHeader() {
-  const { t } = useTranslation('treasury');
-  return (
-    <Box
-      mb="1rem"
-      minW="595px"
-    >
-      <Show above="lg">
-        <Divider
-          my="1rem"
-          variant="darker"
-        />
-      </Show>
-      <HStack px={{ base: '1rem', lg: '1.5rem' }}>
-        <Text
-          w="25%"
-          textStyle="label-small"
-          color="neutral-7"
-        >
-          {t('columnCoins')}
-        </Text>
-        <Text
-          w="30%"
-          textStyle="label-small"
-          color="neutral-7"
-        >
-          {t('columnValue')}
-        </Text>
-        <Text
-          w="45%"
-          textStyle="label-small"
-          color="neutral-7"
-        >
-          {t('columnAllocation')}
-        </Text>
-      </HStack>
-    </Box>
-  );
-}
-
-function CoinRow({
-  safe,
-  totalFiat,
-  asset,
-}: {
-  safe: string;
-  totalFiat: number;
-  asset: TokenDisplayData;
-}) {
-  return (
-    <Flex
-      my="0.5rem"
-      justifyContent="space-between"
-      px={{ base: '1rem', lg: '1.5rem' }}
-      minW="595px"
-      gap="1rem"
-    >
-      <Flex
-        w="25%"
-        alignItems="center"
-        gap="0.5rem"
-      >
-        <Image
-          src={asset.iconUri}
-          fallbackSrc="/images/coin-icon-default.svg"
-          alt={asset.symbol}
-          w="1rem"
-          h="1rem"
-        />
-        <EtherscanLink
-          color="white-0"
-          _hover={{ bg: 'transparent' }}
-          textStyle="body-base"
-          padding={0}
-          borderWidth={0}
-          value={asset.address === zeroAddress ? safe : asset.address}
-          type="token"
-        >
-          {asset.symbol}
-        </EtherscanLink>
-      </Flex>
-      <Flex
-        w="30%"
-        alignItems="flex-start"
-        flexWrap="wrap"
-      >
-        <Text
-          maxWidth="23.8rem"
-          width="100%"
-          isTruncated
-        >
-          <Tooltip
-            label={asset.fullCoinTotal}
-            placement="top-start"
-          >
-            {asset.truncatedCoinTotal}
-          </Tooltip>
-        </Text>
-        <Text
-          textStyle="label-small"
-          color="neutral-7"
-          width="100%"
-        >
-          <Tooltip
-            label={asset.fiatConversion}
-            placement="top-start"
-          >
-            {formatUSD(asset.fiatValue)}
-          </Tooltip>
-        </Text>
-      </Flex>
-      <Flex
-        w="45%"
-        alignItems="flex-start"
-      >
-        <Text>{totalFiat > 0 && formatPercentage(asset.fiatValue, totalFiat)}</Text>
-      </Flex>
-    </Flex>
-  );
-}
-
-function NFTHeader() {
-  const { t } = useTranslation('treasury');
-  return (
-    <Box
-      marginBottom="1rem"
-      minW="595px"
-    >
-      <Divider
-        variant="darker"
-        marginTop="0.75rem"
-        marginBottom="1.5rem"
-      />
-      <Text
-        w="25%"
-        textStyle="label-small"
-        color="neutral-7"
-        px={{ base: '1rem', lg: '1.5rem' }}
-      >
-        {t('columnNFTs')}
-      </Text>
-    </Box>
-  );
-}
-
-function NFTRow({ asset, isLast }: { asset: SafeCollectibleResponse; isLast: boolean }) {
-  const image = asset.imageUri ? asset.imageUri : asset.logoUri;
-  const name = asset.name ? asset.name : asset.tokenName;
-  const id = asset.id.toString();
-
-  return (
-    <HStack
-      marginBottom={isLast ? '0rem' : '1.5rem'}
-      px={{ base: '1rem', lg: '1.5rem' }}
-      minW="595px"
-    >
-      <Flex width="15%">
-        <EtherscanLink
-          type="token"
-          value={asset.address}
-          secondaryValue={id}
-          data-testid="link-nft-image"
-          padding={0}
-          _hover={{ bg: 'transparent' }}
-        >
-          <Image
-            src={image}
-            fallbackSrc="/images/nft-image-default.svg"
-            alt={name}
-            w="3rem"
-            h="3rem"
-            marginRight="0.75rem"
-          />
-        </EtherscanLink>
-      </Flex>
-      <Flex width="30%">
-        <EtherscanLink
-          type="address"
-          value={asset.address}
-          _hover={{ bg: 'transparent' }}
-          color="white-0"
-          textStyle="body-base"
-        >
-          {name}
-        </EtherscanLink>
-      </Flex>
-      <Flex width="25%">
-        <EtherscanLink
-          type="token"
-          value={asset.address}
-          secondaryValue={id}
-          color="white-0"
-          textStyle="body-base"
-          _hover={{ bg: 'transparent' }}
-          maxW="100%"
-        >
-          <Text as="span">{`#${id}`}</Text>
-        </EtherscanLink>
-      </Flex>
-    </HStack>
-  );
-}
+import { useFormatCoins } from '../hooks/useFormatCoins';
+import { CoinHeader, CoinRow } from './AssetCoin';
+import { NFTHeader, NFTRow } from './AssetNFT';
 
 export function Assets() {
   const {
